+++ lib/bones_api.dart
@@ -230,6 +230,7 @@ export 'src/bones_api_config.dart';
 export 'src/bones_api_entity.dart';
 export 'src/bones_api_entity_adapter.dart';
 export 'src/bones_api_entity_adapter_memory.dart';
+export 'src/bones_api_entity_adapter_sql.dart';
 export 'src/bones_api_entity_sql.dart';
 export 'src/bones_api_extension.dart';
 export 'src/bones_api_initializable.dart';